Time: 0 ms (100%), Space: 8.4 MB (71.42%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.c
@@ -1,5 +1,5 @@
 int reverse(int x) {
-    long long int rev_num = 0;
+    long int rev_num = 0;
     
     while (x != 0) {
         int digit = x % 10;
